completed basic ui of edit employee pane
